Add Python-based surrogate for rough surface contact
Add a new contact constitutive law, that does not implement any physics
itself, but uses pybind11 to call a Python script for all evaluations.
Hence, the physical model is implemented / provided by the Python
script.

Calls to pybind11 can lead to visibility warnings of the compiler. This
is a known phenomenon
(https://pybind11.readthedocs.io/en/stable/faq.html#someclass-declared-with-greater-visibility-than-the-type-of-its-field-someclass-member-wattributes).
These warnings are specifically suppressed by specifying the visibility
attribute to "hidden". For convencience, this can be done by a new macro
in 4C: FOUR_C_HIDDEN

For testing, a simple Python model re-implements a linear constitutive
law from C++ in Python. Both give the same results up to machine
precision. An input file tests demonstrates prooper intergration into
the multi-scale framework.

Thanks to @sebproell for his help with the pybind11 integration!