Adapt functionality behind WRITE_FINAL_STATE
When WRITE_FINAL_STATE is true, only the
restart output is written now.
- Documentation is adapted.
- Call prepare_output() needs to be deleted after
structadapter->integrate(), as update has already been called
and the time is already after the MAXTIME of the simulation.
So, potentially relevant history data may have been
destroyed by the update call. Thus, in general, we are not
able to reconstruct a reasonable stress or strain state from this.
- Test case with tested stresses is
fixed by writing the new vtk-output at every time step
(INTERVAL_STEPS to 1). This will trigger
the correct calulation of stresses.
- stresses and strains are no longer written by the
old HDF5 output. So we could already delete
the function determine_stress_strain().
- Special capability in adaptive time stepping to always
write restart output at the last time step is removed. Instead
we set the parameter WRITE_FINAL_STATE (doing the same) to true
in these tests.
- Adapt comment describing test input files
that test the restart from the last time step of the simulation.