Refactor DEM surface energy calculation
Removes the old surface energy handler and replaces it by a function
that calculates the surface energy based on a switch-case. This makes
the code easier to follow and saves some virtual function calls.

The inheritance structure is not necessary as it is not expected that a
user will derive from a base class to implement a new surface energy
calculation. If a new surface energy calculation is needed, the
switch-case can be extended.