Fix bug in calculation of
number of degrees of freedom and number of dimensions for the
preconditioner when a nullspace_node_map has been given.
This did not work, especially in coupled problems as the dof gids
might have an offset and thus no node with that gid was found leaving
the default value in the variables numdf and dimns.

This is now changed. Furthermore, the default values are set to an
unreasonable negative value and it is checked, if it is positive
after the determination procedure.